pyproject.toml
[tool.poetry]
name = "discordpy-template"
version = "0.1.0"
description = ""
authors = ["t4t5u0"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
